requirements.txt
requests
faiss-cpu
sentence-transformers
groq